Split extension lifecycle auth regression test
